add logic controller order and review

+++ src/main/java/dev/noroom/thuvien/model/Book.java
@@ -19,8 +19,8 @@ public class Book {
     @NonNull String description;
     long genreId;
     @NonNull Date releaseDate;
-    int numPages;
-    int sold;
+    long numPages;
+    long sold;
     @Nullable
     String selectedImage;
